pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "djangocms-taxonomy"
description = "Taxonomy support for django CMS"
dynamic = ["version"]
authors = [
    {name = "Fabian Braun", email = "fsbraun@gmx.de"}
]
maintainers = [
    {name = "Django CMS Association and contributors", email = "info@django-cms.org"}
]
license = "BSD-3-Clause"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-cms>=4.1",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Framework :: Django :: 6.0",
    "Framework :: Django CMS",
    "Framework :: Django CMS :: 4.0",
    "Framework :: Django CMS :: 4.1",
    "Framework :: Django CMS :: 5.0",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-django>=4.5.2",
    "pytest-cov>=6.0.0",
    "mypy>=1.3.0",
    "django-stubs>=4.2.3",
    "ruff>=0.0.241",
]

[project.urls]
Homepage = "https://github.com/django-cms/djangocms-taxonomy"
Repository = "https://github.com/django-cms/djangocms-taxonomy"


[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic]
version = {attr = "djangocms_taxonomy.__version__"}

[tool.setuptools.package-data]
djangocms_taxonomy = [
    "static/**",
    "templates/**",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["djangocms_taxonomy*"]

[tool.ruff]
line-length = 119
